% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterflexPlots.R
\name{marginaleffectsPlot}
\alias{marginaleffectsPlot}
\title{Marginal effects plot from interflex}
\usage{
marginaleffectsPlot(Y, D, X, data)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment variable}

\item{X}{Moderating variable}

\item{data}{dataframe}
}
\description{
Marginal effects plot from interflex
}
\examples{
 appdata <- makeAppData(data = mtcars, model = fiti, pred="hp", modx="wt")
 marginaleffectsPlot(Y = "mpg", D = appdata$pred,X = appdata$modx,data = appdata$data)
}
