% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterflexPlots.R
\name{RawPlot}
\alias{RawPlot}
\title{RawPlot from interflex}
\usage{
RawPlot(Y, D, X, Ylabel, Dlabel, Xlabel, data)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment variable}

\item{X}{Moderating variable}

\item{Ylabel}{Label for Y}

\item{Dlabel}{Label for D}

\item{Xlabel}{Label for X}

\item{data}{dataframe}
}
\value{
Plot of Raw data interaction plot
}
\description{
RawPlot from interflex
}
\examples{
appdata <- makeAppData(data = mtcars, model = fiti, pred="hp", modx="wt")
RawPlot(Y = "mpg", D = appdata$pred,X = appdata$modx,Ylabel="mpg", Xlabel=appdata$modx,
Dlabel = appdata$pred,data = appdata$data)
}
